# Conflicts:
#	app/utils/misc.ts
#	package.json

+++ tests/e2e/hook.js
@@ -1,7 +1,7 @@
 /* Copyright (c) 2016-present - TagSpaces UG (Haftungsbeschraenkt). All rights reserved. */
-import { Application } from 'spectron';
 import electronPath from 'electron';
 import pathLib from 'path';
+import fse from 'fs-extra';
 
 // Spectron API https://github.com/electron/spectron
 // Webdriver.io http://webdriver.io/api.html
@@ -28,15 +28,21 @@ export async function clearLocalStorage() {
     global.app.webContents.reload();
   }*/
   if (global.isWeb) {
-    //clearAllURLParams && clears everything in localStorage
-    await global.client.execute(
-      "window.history.pushState('', document.title, window.location.pathname);localStorage.clear();"
-    );
-    // global.client.executeScript('window.localStorage.clear()');
-    // global.client.clearLocalStorage();
-    // window.localStorage.clear();
-    //await global.client.reloadSession();
-    await global.client.refresh();
+    if (isPlaywright) {
+      const windowHandle = await global.client.evaluateHandle(() => window);
+      const title = await global.client.evaluateHandle(() => document.title);
+      windowHandle.history.pushState('', title, windowHandle.location.pathname);
+    } else {
+      //clearAllURLParams && clears everything in localStorage
+      await global.client.execute(
+        "window.history.pushState('', document.title, window.location.pathname);localStorage.clear();"
+      );
+      // global.client.executeScript('window.localStorage.clear()');
+      // global.client.clearLocalStorage();
+      // window.localStorage.clear();
+      //await global.client.reloadSession();
+      await global.client.refresh();
+    }
   } else {
     await global.app.webContents.executeJavaScript(
       "window.history.pushState('', document.title, window.location.pathname);localStorage.clear()"
@@ -48,7 +54,20 @@ export async function clearLocalStorage() {
   // global.app.client.reload(false);
 }
 
-export async function startSpectronApp() {
+export async function copyExtConfig(extconfig = 'extconfig-with-welcome.js') {
+  const srcDir = pathLib.join(__dirname, '..', '..', 'scripts', extconfig);
+  const destDir = pathLib.join(__dirname, '..', '..', 'app', 'extconfig.js');
+  await fse.copy(srcDir, destDir);
+}
+
+export async function removeExtConfig() {
+  await fse.remove(pathLib.join(__dirname, '..', 'app', 'extconfig.js'));
+}
+
+export async function startTestingApp(extconfig) {
+  if (extconfig) {
+    await copyExtConfig(extconfig);
+  }
   const chromeDriverArgs = [
     // '--disable-gpu',
     '--disable-infobars',
@@ -62,58 +81,103 @@ export async function startSpectronApp() {
   }
 
   if (global.isWeb) {
-    //require('scripts/wdio.conf');
-    const webdriverio = require('webdriverio');
-    // https://webdriver.io/docs/configurationfile.html
+    if (global.isPlaywright) {
+      const { webkit, chromium } = require('playwright');
+      global.app = await chromium.launch({
+        headless: global.isHeadlessMode,
+        slowMo: 50
+      }); //browser
 
-    const options = {
-      host: 'localhost', // Use localhost as chrome driver server
-      port: 9515, // "9515" is the port opened by chrome driver.
-      capabilities: {
-        browserName: 'chrome',
-        'goog:chromeOptions': {
-          w3c: true,
-          args: chromeDriverArgs
-        }
-        // pageLoadStrategy: 'normal'
-      },
-      // Warns when a deprecated command is used
-      deprecationWarnings: true,
-      // If you only want to run your tests until a specific amount of tests have failed use
-      // bail (default is 0 - don't bail, run all tests).
-      bail: 0,
-      reporters: ['spec'],
-      /* afterTest: [async function(
-        test,
-        context,
-        { error, result, duration, passed, retries }
-      ) => {
-        await takeScreenshot();
-        await clearLocalStorage();
-      }], */
-      waitforTimeout: 5000,
-      maxInstances: 1,
-      // logLevel: 'debug'
-      logLevel: 'silent',
-      coloredLogs: true
-    };
-    // global.client = browser
-    global.client = await webdriverio.remote(options);
-    // global.client.setTimeout({ 'script': 60000 });
-    /*client = await client
-      .init()
-      .setViewportSize({ width: 1024, height: 768 }, false)
-      .timeouts('script', 6000)
-      /!*
-                 Cannot set 'implicit' timeout because of a bug in webdriverio [1].
-                 [1] https://github.com/webdriverio/webdriverio/issues/974
-                 *!/
-      // .timeouts('implicit', 5000)
-      .timeouts('pageLoad', 30000);*/
-    setWdioImageComparisonService(global.client);
+      global.context = await global.app.newContext({
+        viewport: { width: 1920, height: 1080 }
+      });
 
-    await global.client.url('http://localhost:8000');
+      global.client = await global.context.newPage(); //page
+      await global.client.goto('http://localhost:8000');
+      // await global.client.screenshot({ path: `example.png` });
+      // await global.client.close();
+    } else {
+      //require('scripts/wdio.conf');
+      const webdriverio = require('webdriverio');
+      // https://webdriver.io/docs/configurationfile.html
+
+      const options = {
+        host: 'localhost', // Use localhost as chrome driver server
+        port: 9515, // "9515" is the port opened by chrome driver.
+        capabilities: {
+          browserName: 'chrome',
+          'goog:chromeOptions': {
+            w3c: true,
+            args: chromeDriverArgs
+          }
+          // pageLoadStrategy: 'normal'
+        },
+        // Warns when a deprecated command is used
+        deprecationWarnings: true,
+        // If you only want to run your tests until a specific amount of tests have failed use
+        // bail (default is 0 - don't bail, run all tests).
+        bail: 0,
+        reporters: ['spec'],
+        /* afterTest: [async function(
+          test,
+          context,
+          { error, result, duration, passed, retries }
+        ) => {
+          await takeScreenshot();
+          await clearLocalStorage();
+        }], */
+        waitforTimeout: 5000,
+        maxInstances: 1,
+        // logLevel: 'debug'
+        logLevel: 'silent',
+        coloredLogs: true
+      };
+      // global.client = browser
+      global.client = await webdriverio.remote(options);
+      // global.client.setTimeout({ 'script': 60000 });
+      /*client = await client
+        .init()
+        .setViewportSize({ width: 1024, height: 768 }, false)
+        .timeouts('script', 6000)
+        /!*
+                   Cannot set 'implicit' timeout because of a bug in webdriverio [1].
+                   [1] https://github.com/webdriverio/webdriverio/issues/974
+                   *!/
+        // .timeouts('implicit', 5000)
+        .timeouts('pageLoad', 30000);*/
+      setWdioImageComparisonService(global.client);
+
+      await global.client.url('http://localhost:8000');
+    }
+  } else if (global.isPlaywright) {
+    const { _electron: electron } = require('playwright');
+    // Launch Electron app.
+    global.app = await electron.launch({
+      args: [pathLib.join(__dirname, '..', '..', 'app', 'main.prod.js')]
+    });
+
+    // Get the first window that the app opens, wait if necessary.
+    global.client = await global.app.firstWindow();
+    await global.client.waitForLoadState('load'); //'domcontentloaded'); //'networkidle');
+    // await global.client.bringToFront();
+
+    // Print the title.
+    // console.log(await global.client.title());
+
+    if (process.env.SHOW_CONSOLE) {
+      // Direct Electron console to Node terminal.
+      global.client.on('console', console.debug);
+    }
+    // Click button.
+    /*await global.client.click('[data-tid=location_supported-filestypes]');
+    // Capture a screenshot.
+    await global.client.screenshot({
+      path: pathLib.join(__dirname, 'intro.png')
+    });*/
+    // Exit app.
+    // await global.app.close();
   } else {
+    const { Application } = require('spectron');
     global.app = new Application({
       path: electronPath,
       args: [pathLib.join(__dirname, '..', '..', 'app')],
@@ -150,18 +214,22 @@ function setWdioImageComparisonService(browser) {
 }
 
 export async function stopSpectronApp() {
-  if (global.app && global.app.isRunning()) {
+  if (global.isPlaywright && global.app) {
+    await global.app.close();
+  } else if (global.isWeb) {
+    await global.client.closeWindow();
+    // await global.client.end();
+  } else if (global.app && global.app.isRunning()) {
     // await clearLocalStorage();
     return global.app.stop();
   }
-  if (global.isWeb) {
-    await global.client.closeWindow();
-    // await global.client.end();
-  }
 }
 
-export function testDataRefresh() {
+export async function testDataRefresh() {
   const fse = require('fs-extra');
+  //const gracefulFs = require('graceful-fs')
+  //gracefulFs.gracefulify(fse);
+
   const src = pathLib.join(
     __dirname,
     '..',
@@ -174,6 +242,9 @@ export function testDataRefresh() {
   let newPath = pathLib.join(dst, pathLib.basename(src));
   fse.emptyDirSync(newPath);
   fse.copySync(src, newPath, { overwrite: true });
+  /*if (global.isElectron && global.client) {
+    await global.client.waitForTimeout(1000);
+  }*/
 }
 
 export async function takeScreenshot(name = expect.getState().currentTestName) {
